added spotify auth route to playlists
